# contracts/openapi/base.yaml
openapi: 3.0.3
info:
  title: Spider API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/projects:
    get:
      summary: List projects
      parameters:
        - in: query
          name: workspaceId
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProjectList"
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "200":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProject"
  /api/v1/projects/{projectId}:
    get:
      summary: Get project detail
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProject"
    put:
      summary: Update project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseProject"
    delete:
      summary: Delete project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseDelete"
...
  /api/v1/requirement-handoffs:
    get:
      summary: List published requirement handoffs
      parameters:
        - in: query
          name: projectId
          required: false
          schema:
            type: string
            format: uuid
        - in: query
          name: targetRole
          required: false
          schema:
            type: string
            enum: [product, backend, frontend, test]
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [draft, in_review, published, accepted, in_development, in_testing, done, rejected]
      responses:
        "200":
          description: Requirement handoff list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRequirementHandoffList"
  /api/v1/requirement-handoffs/{handoffId}/accept:
    post:
      summary: Accept a published requirement handoff
      parameters:
        - in: path
          name: handoffId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Requirement handoff accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRequirementHandoff"
  /api/v1/requirement-handoffs/{handoffId}/transition:
    post:
      summary: Transition requirement handoff state
      parameters:
        - in: path
          name: handoffId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequirementHandoffRequest"
      responses:
        "200":
          description: Requirement handoff transitioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRequirementHandoff"
  /api/v1/requirement-handoffs/{handoffId}/tasks/ai-generate:
    post:
      summary: Generate role task plan by AI for requirement handoff
      parameters:
        - in: path
          name: handoffId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: AI task plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAiTaskPlan"
  /api/v1/requirement-handoffs/{handoffId}/tasks:
    get:
      summary: List requirement tasks under handoff
      parameters:
        - in: path
          name: handoffId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Requirement task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRequirementTaskList"
  /api/v1/requirement-handoffs/tasks/{taskId}/claim:
    post:
      summary: Claim requirement task to current actor
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Requirement task claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRequirementTask"
  /api/v1/requirement-handoffs/tasks/{taskId}/status:
    put:
      summary: Update requirement task status
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequirementTaskStatusRequest"
      responses:
        "200":
          description: Requirement task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRequirementTask"
  /api/v1/tasks:
    get:
      summary: List tasks for task center views
      parameters:
        - in: query
          name: view
          required: false
          schema:
            type: string
            enum: [pool, mine, all]
            default: pool
        - in: query
          name: projectId
          required: false
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [todo, in_progress, done, blocked]
        - in: query
          name: role
          required: false
          schema:
            type: string
            enum: [product, backend, frontend, test]
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Task center page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTaskCenterPage"
  /api/v1/tasks/{taskId}:
    get:
      summary: Get task detail with requirement handoff context
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Requirement task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRequirementTaskDetail"
...
components:
  schemas:
...
    ApiResponseProject:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/Project"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    ApiResponseRequirementHandoff:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/RequirementHandoff"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    ApiResponseAiTaskPlan:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/AiTaskPlan"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    ApiResponseProjectList:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    ApiResponseRequirementHandoffList:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/RequirementHandoff"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    ApiResponseRequirementTask:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/RequirementTask"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseRequirementTaskDetail:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/RequirementTaskDetail"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseRequirementTaskList:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/RequirementTask"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseTaskCenterPage:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          $ref: "#/components/schemas/TaskCenterPage"
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    ApiResponseDelete:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
          additionalProperties:
            type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RequirementHandoff:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        sourceSessionId:
          type: string
          format: uuid
        version:
          type: integer
        title:
          type: string
        requirementSummary:
          type: string
        acceptanceCriteria:
          type: string
        impactScope:
          type: string
        priority:
          type: string
          enum: [P0, P1, P2, P3]
        targetRole:
          type: string
          enum: [product, backend, frontend, test]
        status:
          type: string
          enum: [draft, in_review, published, accepted, in_development, in_testing, done, rejected]
        publishedBy:
          type: string
        publishedAt:
          type: string
          format: date-time
        acceptedBy:
          type: string
        acceptedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RequirementTask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        handoffId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        role:
          type: string
          enum: [product, backend, frontend, test]
        title:
          type: string
        titleZh:
          type: string
        titleEn:
          type: string
        description:
          type: string
        descriptionZh:
          type: string
        descriptionEn:
          type: string
        estimateHours:
          type: integer
        status:
          type: string
          enum: [todo, in_progress, done, blocked]
        assignee:
          type: string
        source:
          type: string
          enum: [ai, manual]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RequirementTaskDetail:
      allOf:
        - $ref: "#/components/schemas/RequirementTask"
        - type: object
          properties:
            projectName:
              type: string
            handoffTitle:
              type: string
            handoffRequirementSummary:
              type: string
            handoffAcceptanceCriteria:
              type: string
            handoffImpactScope:
              type: string
            handoffPriority:
              type: string
              enum: [P0, P1, P2, P3]
            handoffTargetRole:
              type: string
              enum: [product, backend, frontend, test]
            handoffStatus:
              type: string
              enum: [draft, in_review, published, accepted, in_development, in_testing, done, rejected]
    TaskCenterPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RequirementTask"
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
        hasNext:
          type: boolean
...
    AiTaskPlan:
      type: object
      properties:
        handoffId:
          type: string
          format: uuid
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/RequirementTask"
        impactedFiles:
          type: array
          items:
            type: string
        riskHints:
          type: array
          items:
            type: string
        testHints:
          type: array
          items:
            type: string
        provider:
          type: string
        model:
          type: string
        rationale:
          type: string
...
    CreateProjectRequest:
      type: object
      required: [workspaceId, name, status]
      properties:
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
...
    UpdateProjectRequest:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
...
    TransitionRequirementHandoffRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [submit_review, publish, accept, start_dev, start_test, complete, reject, reopen]
    UpdateRequirementTaskStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [todo, in_progress, done, blocked]
...